requirement.txt
pytest~=7.4.4
pytest-html
pytest-xdist
pytest-ordering
pytest-rerunfailures
pytest-base-url
allure-pytest
PyYAML~=6.0.1
redis~=5.0.4
colorlog~=6.8.2
requests~=2.32.2
httprunner~=4.3.5
jsonpath~=0.82.2